Move transaction-related golden tests from cardano-ledger-alonzo-test to cardano-ledger-alonzo
Consolidates golden tests from the separate `cardano-ledger-alonzo-test`
package into the main `cardano-ledger-alonzo` impl package, following the
same pattern already completed in other eras (allegra, mary, conway, babbage).
