+++ eras/alonzo/test-suite/cardano-ledger-alonzo-test.cabal
@@ -86,7 +86,6 @@ test-suite cardano-ledger-alonzo-test
     -rtsopts
 
   build-depends:
-    aeson >=2,
     base,
     base16-bytestring,
     bytestring,
@@ -94,15 +93,12 @@ test-suite cardano-ledger-alonzo-test
     cardano-ledger-alonzo-test,
     cardano-ledger-binary:{cardano-ledger-binary, testlib},
     cardano-ledger-core:{cardano-ledger-core, testlib},
-    cardano-ledger-mary,
     cardano-ledger-shelley:{cardano-ledger-shelley, testlib},
-    cardano-ledger-shelley-ma-test,
     cardano-ledger-shelley-test,
     cardano-protocol-tpraos:{cardano-protocol-tpraos, testlib},
     cardano-slotting,
     cardano-strict-containers,
     containers,
     microlens,
-    plutus-ledger-api,
     small-steps:{small-steps, testlib} >=1.1,
     time,
